requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Werkzeug==2.3.7
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
Pillow==10.0.0
